requirements.txt
openai
pdfplumber
scipy
spacy
python-dotenv
psycopg2-binary
